OKE-617
git-svn-id: https://submarine.computas.com/sublima/trunk@2850 8fc9d99b-3020-0410-9de1-dfa816a989ce

+++ blocks/sublima-app/src/main/java/com/computas/sublima/app/controller/admin/TopicController.java
@@ -383,7 +383,7 @@ public class TopicController implements StatelessAppleController {
       updateSuccess = sparulDispatcher.query(insertNewTopicString);
 
       for (String oldurl : req.getCocoonRequest().getParameterValues("skos:Concept")) {
-        if (!"".equals(oldurl) || oldurl != null) {
+        if (!"".equals(oldurl) && oldurl != null) {
           //String sparulQuery = "MODIFY\nDELETE { ?s ?p <" + oldurl + "> }\nINSERT { ?s ?p <" + uri + "> }\nWHERE { ?s ?p <" + oldurl + "> }\n";
           String sparulQuery = "INSERT INTO <" + SettingsService.getProperty("sublima.basegraph") + ">{ ?s ?p <" + uri + "> }\nWHERE { ?s ?p <" + oldurl + "> }\n";
           String sparulDelete = "DELETE FROM <" + SettingsService.getProperty("sublima.basegraph") + ">{ ?s ?p <" + oldurl + "> }\nWHERE { ?s ?p <" + oldurl + "> }\n";
